+++ gridcomps/History/MAPL_HistoryGridComp.F90
@@ -53,11 +53,11 @@
   use pFIO_ConstantsMod
   use HistoryTrajectoryMod
   use StationSamplerMod
-  use MaskSamplerGeosatMod
+  use MaskSamplerMod
   use MAPL_StringTemplate
   use regex_module
   use MAPL_TimeUtilsMod, only: is_valid_time, is_valid_date, MAPL_UndefInt
-  use gFTL_StringStringMap
+  use gFTL2_StringStringMap
   !use ESMF_CFIOMOD
   use MAPL_EpochSwathMod
 
@@ -578,7 +578,7 @@ contains
 
              nlist = nlist + 1
              allocate( list(nlist), _STAT )
-             list(1:nlist-1)=IntState%list
+             if (nlist > 1) list(1:nlist-1)=IntState%list
              list(nlist)%collection = tmpstring
              list(nlist)%filename = list(nlist)%collection
              deallocate(IntState%list)
@@ -2493,8 +2493,16 @@ ENDDO PARSER
              IntState%stampoffset(n) = list(n)%trajectory%epoch_frequency
           elseif (list(n)%sampler_spec == 'mask') then
              call MAPL_TimerOn(GENSTATE,"mask_init")
-             list(n)%mask_sampler = MaskSamplerGeosat(cfg,string,clock,genstate=GENSTATE,_RC)
-             call list(n)%mask_sampler%initialize(items=list(n)%items,bundle=list(n)%bundle,timeinfo=list(n)%timeInfo,vdata=list(n)%vdata,_RC)
+             global_attributes = list(n)%global_atts%define_collection_attributes(_RC)
+             list(n)%mask_sampler = MaskSampler(cfg,string,clock,genstate=GENSTATE,_RC)
+             ! initialize : create grid / metadata
+             call list(n)%mask_sampler%set_param(oClients=o_Clients)
+             call list(n)%mask_sampler%set_param(itemOrder=intState%fileOrderAlphabetical,_RC)
+             call list(n)%mask_sampler%initialize(list(n)%duration,list(n)%frequency,items=list(n)%items,&
+                  bundle=list(n)%bundle,timeinfo=list(n)%timeInfo,vdata=list(n)%vdata,global_attributes=global_attributes,_RC)
+
+             collection_id = o_Clients%add_data_collection(list(n)%mask_sampler%metadata, mode = create_mode)
+             call list(n)%mask_sampler%set_param(write_collection_id=collection_id)
              call MAPL_TimerOff(GENSTATE,"mask_init")
           elseif (list(n)%sampler_spec == 'station') then
              list(n)%station_sampler = StationSampler (list(n)%bundle, trim(list(n)%stationIdFile), nskip_line=list(n)%stationSkipLine, genstate=GENSTATE, _RC)
@@ -3549,7 +3557,7 @@ ENDDO PARSER
             if (intState%allow_overwrite) create_mode = PFIO_CLOBBER
             ! add time to items
             ! true metadata comes here from mGriddedIO%metadata
-            ! the mGriddedIO below only touches metadata, collection_id etc., it is safe.
+            ! the list(n)%mGriddedIO below only touches metadata, collection_id etc.
             !
             if (.NOT. list(n)%xsampler%have_initalized) then
                list(n)%xsampler%have_initalized = .true.
@@ -3642,13 +3650,18 @@ ENDDO PARSER
                list(n)%unit = -1
             end if
          elseif (list(n)%sampler_spec == 'mask') then
-            if (list(n)%unit.eq.0) then
-               call lgr%debug('%a %a',&
-                    "Mask_data output to new file:",trim(filename(n)))
-               call list(n)%mask_sampler%close_file_handle(_RC)
-               call list(n)%mask_sampler%create_file_handle(filename(n),_RC)
-               list(n)%currentFile = filename(n)
-               list(n)%unit = -1
+            if( list(n)%unit.eq.0 ) then
+               if (list(n)%format == 'CFIO') then
+                  if (.not.intState%allow_overwrite) then
+                     inquire (file=trim(filename(n)),exist=file_exists)
+                     _ASSERT(.not.file_exists,trim(filename(n))//" being created for History output already exists")
+                  end if
+!!                  call list(n)%mask_sampler%modifyTime(oClients=o_Clients,_RC)
+                  list(n)%currentFile = filename(n)
+                  list(n)%unit = -1
+               else
+                  list(n)%unit = GETFILE( trim(filename(n)),all_pes=.true.)
+               end if
             end if
          else
             if( list(n)%unit.eq.0 ) then
@@ -3795,10 +3808,13 @@ ENDDO PARSER
          elseif (list(n)%sampler_spec == 'mask') then
             call ESMF_ClockGet(clock,currTime=current_time,_RC)
             call MAPL_TimerOn(GENSTATE,"Mask_append")
-            call list(n)%mask_sampler%append_file(current_time,_RC)
+            if (list(n)%unit < 0) then    ! CFIO
+               call list(n)%mask_sampler%regrid_append_file(current_time,&
+                    list(n)%currentFile,oClients=o_Clients,_RC)
+               call lgr%debug('%a %a', 'mask sampler list(n)%currentFile: ', trim(list(n)%currentFile))
+            end if
             call MAPL_TimerOff(GENSTATE,"Mask_append")
          endif
-
 
       endif OUTTIME
 
@@ -3820,6 +3836,7 @@ ENDDO PARSER
       call o_Clients%post_wait()
    endif
    call MAPL_TimerOff(GENSTATE,"Done Wait")
+
 
   ! destroy ogrid/RH/acc_bundle, regenerate them
   ! swath only
@@ -3927,6 +3944,12 @@ ENDDO PARSER
     IntState => wrap%ptr
     list => IntState%list
     nlist = size(list)
+
+    do n=1,nlist
+       if (list(n)%sampler_spec == 'mask') then
+          call list(n)%mask_sampler%finalize(_RC)
+       end if
+    end do
 
 ! Close UNITs of GEOSgcm History Data
 ! -----------------------------------